Updated Backend + AI!
This was a pretty long one. There was  a time when this whole Backend thing had 1000+ Problems. Now, it's 0. Yes. But, AI isn't working yet. It has to be implemented carefully. Some invisible bugs are still breaking the whole thing. This commit is basically to save the progress I have made.